[ALPS07993152] fix coverity
0. copy from 7386270 (sync repo)
1. use enum IMGSENSOR_MODE_MAX for table size
2. fix int overflow @ adaptor-subdrv-ctrl.c

MTK-Commit-Id: c5500f95c1bac8e7c5f4a3e5830b28bf109ac193

CR-Id: ALPS07993152
Feature: [Android Default] Camera Application Basic Functions

+++ imgsensor/src-v4l2/adaptor-subdrv-ctrl.c
@@ -2056,7 +2056,7 @@ void get_min_shutter_by_scenario(struct subdrv_ctx *ctx,
 		enum SENSOR_SCENARIO_ID_ENUM scenario_id,
 		u64 *min_shutter, u64 *exposure_step)
 {
-	u32 ratio = 1;
+	u64 ratio = 1;
 
 	if (scenario_id >= ctx->s_ctx.sensor_mode_num) {
 		DRV_LOG(ctx, "invalid cur_sid:%u, mode_num:%u set default\n",
